requirements.txt
python-dotenv~=0.19.2
pytelegrambotapi==4.4.0